Update patch set 1
Patch Set 1:

(4 comments)

I have put in a few thoughts.
Mainly, the code needs reformatting according go tool specifications.
And os calls and native library calls don't need to be mocked.
It is fine to mock the DB calls. If you are not using an interface to abstract the database, then this is fine. However, if you are indeed using an interface for the DB, then just implement that interface in your test and use that similar to how we do it in k8splugin.

Patch-set: 1
Label: Code-Review=0