Reliable volume for minio server and e2e support
1. Add E2E test for MinIO Server side, since the image upload to
MinIO Server and we need to check the actual size of the object
in MinIO Server, after PATCH operations, check the object size
with REST parameters with image size.

2. Fix an issue for MinIO Service, in the install.sh of MinIO
Server, the minio-service.yaml need to be created and then we
can get the Endpoint IP for Minio client operations.

3. Reliable Volume implementation, which use Ceph CSI to provide
Ceph RBD Volume to export by MinIO. But this feature is NOT enabled
yet, and NO impact for current local volume, since there are some
other changes since I saw REST API Agent service also use the volume
with MinIO, this feature can be enable in next release.

4. Add uninstall script to clean the Storageclass/PV/PVC and other
resources in MinIO Server deloyment.

Change-Id: I17250be6ee1d64b98ad789673621f2f2e0009c3e
Signed-off-by: Chen, Tingjie <tingjie.chen@intel.com>