Use username and password from "os" in nodes.json
This removes COMPUTE_NODE_PASSWORD from the environment.  A check is
left in place to ensure that any existing users of
COMPUTE_NODE_PASSWORD will be warned that the nodes.json value is what
is used now.

Also, COMPUTE_NODE_FQDN is removed from common.sh.  It was/is
overwritten immediately before use.

Note: the "os" object does not exist in the equivalent JSON file for
the VM deployment.  The default username and password continues to be
used in the VM case.

Issue-ID: ICN-497
Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>
Change-Id: I3cab61cf610d7ed334ff0043cb5f7f4ed442662f