Make ipa-downloader error fail the build
Running the image in the background and then issuing docker wait fails
to detect a failed return code from the container as docker wait
prints the return code of the container and exits with 0.

Check explicitly that the ironic-python-agent files the httpd
container needs are downloaded.

This change also removes extraneous whitespace from the generated
ironic.env.

Issue-ID: ICN-499
Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>
Change-Id: Icc0f7ce5b19c47008c37d59c3f029809c68b16cc