Add explicit network configuration to nodes.json
A "net" section is added to nodes.json to provide the value of the
networkData field to the created BareMetalHost resource.  This allows
static configuration of a given host's interface, for example.

Due to limitations with the networkData field, all interfaces must now
be explicitly listed.  Previously, dhclient was run against all
interfaces on the host and gateway and DNS information was provided
via the PROVIDER_NETWORK variables in user_config.sh.  Now that
information must be provided in nodes.json.

Issue-ID: ICN-575
Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>
Change-Id: Icbf29856a0e1209bd1bcfe7da563f8e6cc1057cd