Reduce machine, cluster charts to single instances
The prior versions of the chart allow defining multiple instances of
machines and clusters to be defined in a single release of the chart.
This change now requires that each instance be defined as a unique
release.  While slightly more verbose intially, this simplifies
documentation and allows default values to be provided in a sensible
way.

Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>
Change-Id: I9cd9694cf5de59439b393d5c9d271783d8c7ad54