Add Istio to vm and pod11 sites
This requires generating cluster-specific certificates to be deployed
before Istio is deployed, so split out the sites into separate
deployment and cluster directories.  The deployment directory contains
the site.yaml and the cluster directory contains the cluster
specifics.

As part of this, refactor common site build-source functionality into
site.sh.

Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>
Change-Id: I434e261a67938c2d4f8c7ee7bd4b7154f6325c5f